moved some cheese fixed some minor todo's
