+++ src/Elasticsearch.Net/Configuration/IConnectionConfiguration.cs
