+++ src/Elasticsearch.Net/Connection/Audit/Audit.cs
@@ -1,0 +1,19 @@
+using System;
+
+namespace Elasticsearch.Net.Connection
+{
+	public class Audit
+	{
+		public AuditEvent Event { get; internal set; }
+		public DateTime Started { get; }
+		public DateTime Ended { get; internal set; }
+		public Node Node { get; internal set; }
+		public string Path { get; internal set; }
+
+		public Audit(AuditEvent type, DateTime occured)
+		{
+			this.Event = type;
+			this.Started = occured;
+		}
+	}
+}
