+++ src/Elasticsearch.Net/Connection/Audit/AuditEvent.cs
@@ -1,0 +1,14 @@
+namespace Elasticsearch.Net.Connection
+{
+	public enum AuditEvent
+	{
+		SniffOnStartup,
+		SniffOnFail,
+		SniffOnStaleCluster,
+		SniffFail,
+		SniffSuccess,
+		Ping,
+		BadResponse,
+		HealhyResponse
+	}
+}
