--- src/Elasticsearch.Net/Connection/Audit/RequestType.cs
@@ -1,9 +1,0 @@
-namespace Elasticsearch.Net
-{
-	public enum RequestType
-	{
-		ElasticsearchCall,
-		Ping,
-		Sniff
-	}
-}
