+++ src/Elasticsearch.Net/Connection/Pipeline/RequestPipeline.cs
@@ -11,60 +11,6 @@ using System.Threading.Tasks;
 
 namespace Elasticsearch.Net.Connection
 {
-	public enum AuditEvent
-	{
-		SniffOnStartup,
-		SniffOnFail,
-		SniffOnStaleCluster,
-		SniffFail,
-		SniffSuccess,
-		Ping,
-		BadResponse,
-		HealhyResponse
-	}
-
-	public class Audit
-	{
-		public AuditEvent Event { get; internal set; }
-		public DateTime Started { get; }
-		public DateTime Ended { get; internal set; }
-		public Node Node { get; internal set; }
-		public string Path { get; internal set; }
-
-		public Audit(AuditEvent type, DateTime occured)
-		{
-			this.Event = type;
-			this.Started = occured;
-		}
-	}
-
-	internal class Auditable : IDisposable
-	{
-		readonly List<Audit> _auditTrail;
-		readonly IDateTimeProvider _dateTimeProvider;
-		private readonly DateTime _started;
-		private readonly Audit _audit;
-
-		public Node Node { set { this._audit.Node = value; } }
-		public AuditEvent Event { set { this._audit.Event = value; } }
-		public string Path { set { this._audit.Path = value; } }
-
-		public Auditable(AuditEvent type, List<Audit> auditTrail, IDateTimeProvider dateTimeProvider)
-		{
-			this._dateTimeProvider = dateTimeProvider;
-			this._auditTrail = auditTrail;
-			this._started = _dateTimeProvider.Now();
-			this._audit = new Audit(type, this._started);
-			this._auditTrail.Add(this._audit);
-		}
-
-		public void Dispose()
-		{
-			this._audit.Ended = this._dateTimeProvider.Now();
-		}
-	}
-
-
 
 	public class RequestPipeline : IRequestPipeline
 	{
@@ -410,7 +356,5 @@ namespace Elasticsearch.Net.Connection
 		}
 
 		private Auditable Audit(AuditEvent type) => new Auditable(type, this.AuditTrail, this._dateTimeProvider);
-
-
 	}
 }
