+++ src/Elasticsearch.Net/Connection/SpecialResponses/DynamicDictionary.cs
