+++ src/Elasticsearch.Net/Connection/SpecialResponses/ElasticsearchDynamicValue.cs
