+++ src/Elasticsearch.Net/Connection/SpecialResponses/VoidResponse.cs
